updating exception cypress tests under shared lists folder
